Weighted Vertex LSQ is working.
That's what I get for copying lines of code and not being carefull...
will I ever learn...
Works for the linear test function.  We will have to see how the vertex gradient works on non-cartesian meshes.  I may want to add an L2 weighting for the average.  I think that's what the paper does.