Updated method for computing interface gradient at domain boundaries.
Added a method similar to interface boundaries that uses uses a finite difference to calculate the normal component of the interface gradient.  Instead of using the cell center of a right hand ghost cell the face center is used to construct the term ds.  Note: the face value is never explicitly calculated instead it is solved for as a function of q1 and qb and that is used for the computation.