Overhauled AMG to allow for F and W cycles.
Rather than perform a single AMG V-cycle per outer iteration succesive AMG cycles are performed.  For each cycle, either a V, W, or F type cycle is used, with a prescribed number of pre and post sweeps.  In some very simple testing of my unit case the number of fine (level 0) GS sweeps decreased from ~10-12 to 2-6. Still to do: implement Ruge-Stuben coarsenin as opposed to whatever it is that I came up with, and build an AMG structure that stores the restricted matrices RAP as well as the individual restriction (R) and prolongation (P) matrices so that they do not need to be re-computed each amg cycle.