Adjust boundary procedure to allow reconstruction of ghost values to match LHS reconstruction.
The ghost values are computed before reconstruction.  Moving that out of the flux calculation would be a bit of a pain, so as a short term fix we're reconstructing the ghost values.  Essentially applying a specific gradient (often a function of the boundary cell) to ensure the RHS properly matches the LHS.